requirements.txt
streamlit
pandas
unidecode
matplotlib
seaborn
qrcode
Pillow
openpyxl